requirements.txt
livekit-agents>=0.5.0
livekit>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
asyncio-mqtt>=0.16.0
numpy>=1.24.0
torch>=2.0.0
torchaudio>=2.0.0
onnxruntime>=1.15.0
soundfile>=0.12.0
librosa>=0.10.0
websockets>=11.0.0
aiohttp>=3.8.0
pydantic>=2.0.0
typing-extensions>=4.5.0
Flask>=2.0.0
twilio>=8.0.0 